Refactor duke to Brian
