+++ src/main/java/brian/Launcher.java
@@ -1,0 +1,9 @@
+package brian;
+
+import javafx.application.Application;
+
+public class Launcher {
+    public static void main(String[] args) {
+        Application.launch(Main.class, args);
+    }
+}
