+++ src/main/java/brian/command/ByeCommand.java
@@ -1,0 +1,22 @@
+package brian.command;
+
+import brian.storage.Storage;
+import brian.task.TaskList;
+import brian.ui.TextUi;
+import brian.utils.DukeException;
+
+public class ByeCommand extends Command {
+
+    @Override
+    public void execute(TaskList tasks, TextUi ui, Storage storage) throws DukeException {
+        assert ui != null;
+        assert storage != null;
+        storage.save(tasks);
+        ui.showBye();
+    }
+
+    @Override
+    public boolean isExit() {
+        return true;
+    }
+}
