+++ src/main/java/brian/task/Todo.java
@@ -1,0 +1,26 @@
+package brian.task;
+
+public class Todo extends Task {
+    public Todo(String task) {
+        super(task);
+    }
+
+    public Todo(String task, TodoState todoState) {
+        super(task, todoState);
+    }
+
+    @Override
+    public String toString() {
+        return "[T] " + super.toString();
+    }
+
+    @Override
+    public String toFileString() {
+        return "T | " + (todoState == TodoState.DONE ? "1" : "0") + " | " + task;
+    }
+
+    @Override
+    public int getOrder() {
+        return 2;
+    }
+}
