+++ src/main/java/brian/task/TodoState.java
@@ -1,0 +1,5 @@
+package brian.task;
+
+public enum TodoState {
+    UNDONE, DONE
+}
