+++ src/main/java/brian/utils/DukeException.java
@@ -1,0 +1,7 @@
+package brian.utils;
+
+public class DukeException extends Exception {
+    public DukeException(String message) {
+        super(message);
+    }
+}
