--- src/main/java/duke/Main.java
@@ -1,35 +1,0 @@
-package duke;
-
-import java.io.IOException;
-
-import duke.gui.MainWindow;
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Scene;
-import javafx.scene.layout.AnchorPane;
-import javafx.stage.Stage;
-
-public class Main extends Application {
-
-    private Duke duke;
-
-    public Main() throws IOException {
-        this.duke = new Duke();
-    }
-
-    @Override
-    public void start(Stage primaryStage) throws Exception {
-        try {
-
-            FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource("/view/MainWindow.fxml"));
-            AnchorPane ap = fxmlLoader.load();
-            Scene scene = new Scene(ap);
-            primaryStage.setScene(scene);
-
-            fxmlLoader.<MainWindow>getController().setDuke(duke);
-            primaryStage.show();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-}
