--- src/main/java/duke/command/ByeCommand.java
@@ -1,22 +1,0 @@
-package duke.command;
-
-import duke.storage.Storage;
-import duke.task.TaskList;
-import duke.ui.TextUi;
-import duke.utils.DukeException;
-
-public class ByeCommand extends Command {
-
-    @Override
-    public void execute(TaskList tasks, TextUi ui, Storage storage) throws DukeException {
-        assert ui != null;
-        assert storage != null;
-        storage.save(tasks);
-        ui.showBye();
-    }
-
-    @Override
-    public boolean isExit() {
-        return true;
-    }
-}
