--- src/main/java/duke/command/UnmarkCommand.java
@@ -1,22 +1,0 @@
-package duke.command;
-
-import duke.storage.Storage;
-import duke.task.TaskList;
-import duke.ui.TextUi;
-
-public class UnmarkCommand extends Command {
-    private final int taskNumber;
-
-    public UnmarkCommand(int taskNumber) {
-        this.taskNumber = taskNumber;
-    }
-
-    @Override
-    public void execute(TaskList tasks, TextUi ui, Storage storage) {
-        assert ui != null;
-        assert tasks != null;
-
-        tasks.unmarkAsDone(taskNumber);
-        ui.showUnmarked(tasks.getTask(taskNumber));
-    }
-}
