--- src/main/java/duke/task/TodoState.java
@@ -1,5 +1,0 @@
-package duke.task;
-
-public enum TodoState {
-    UNDONE, DONE
-}
