--- src/main/java/duke/utils/DukeException.java
@@ -1,7 +1,0 @@
-package duke.utils;
-
-public class DukeException extends Exception {
-    public DukeException(String message) {
-        super(message);
-    }
-}
